move models to config package
