+++ Engine/Programs/DirkBuildTool/Source/module/header.go
@@ -1,7 +1,7 @@
 package module
 
 import (
-	"DirkBuildTool/models"
+	"DirkBuildTool/config"
 )
 
 type HeaderModule struct {
@@ -12,7 +12,7 @@ type HeaderModule struct {
 
 func (m *HeaderModule) GetName() string            { return m.Name }
 func (m *HeaderModule) GetIncludeDirs() []string   { return m.IncludeDirs }
-func (m *HeaderModule) GetDefines() models.Defines { return nil }
+func (m *HeaderModule) GetDefines() config.Defines { return nil }
 func (m *HeaderModule) GetLibs() []string          { return m.External }
 
 func (m *HeaderModule) Build() error  { return nil }
